requirements.txt
google-genai
firebase-admin
python-dotenv
playwright
requests
pillow